New approach for job container file system aspects
This change reworks the way that the file system aspects function for job containers.
This is a pretty radical difference, and adds in some new Windows APIs to
accomplish the job.

The gist of things, is instead of mounting the wcifs volume containing the unioned
view of the containers layers on the host at a certain path, and then treating this
as the "C:\" of the container, we now skip mounting the volume and now overlay the
container image with the host. This might sound odd, but it's possible for two reasons.

1. There is a file system minifilter in Windows that allows you to bind a filesystem
namespace to another path, so you see whatever was at the first path at the second.
Essentially a bind mount in Linux.
2. This filter allows you to create these bindings at a per silo level. So only
processes in that silo will see these bindings, so it will appear to be a unique
file system layout in each.

Before this change, the containers were just comprised of regular job objects so
we need to first upgrade the job to a silo, and then we start employing these bindings.

* Add bindings for the Bind Filter from bindfltapi.dll
* Define a silo flag to be able to upgrade the containers job object to a silo
* Rework the pod sandbox container for job containers. Previously we faked out
the pod sandbox container, as an optimization. But if we launch a normal Windows
Server Container as the pod sandbox always, then we can take advantage of the
network compartment that will be created, so we don't always have to use the
hosts network.
* Rework directory mounts for job containers to also make use of the Bind Filter.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>