Remove filepath.Clean usage + touchup job containers path tests
This change does a couple of things for the path resolution logic.

1. I was using filepath.Clean to remove any initial '.' paths in the command line
to handle relative paths with a dot. However it changes all forward slashes
to backslashes, which hinders command lines like the following:

`"cmd /c ping 127.0.0.1"`

as the /c argument will be altered. Windows Server containers don't handle
relative paths with a dot, so it doesn't seem necessary to support them for
job containers either, so just get rid of the call entirely.

2. Remove empty spaces off the end of the cmdline if it's quoted. There was
an empty space in this case as I was using strings.Join to join the arguments
after the quoted element. This had no effects on actual usage/functionality,
but to compare commandlines for tests, this made things easier.

3. When replacing instances of %CONTAINER_SANDBOX_MOUNT_POINT% in the commandline,
the path of the volume we were replacing it with had a trailing forward slash, so
you'd end up with C:\C\12345678abcdefg\\\mybinary.exe (two forward slashes). This
also didn't have much of an effect on anything, but just to clean things up.

4. Lastly, this change also refactors the job container path tests as they
were a bit unwieldy and were due for the t.Run treatment. This moves the
tests to being run via t.Run and a slice of configs to test different path,
working directory, env matrixes. Also adds some new test cases to try out.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>