Only fetch Hcn feature set once
When looking at kube-proxy recently, I noticed that while the sync.Once to only
log the feature set and version is helpful, we'd still call into HNS to get
the version every single time, as well as unneccessarily checking the version
ranges for every feature anytime any of the `hcn.IsThingSupported` functions
are called.

This change moves all of the code for `hcn.GetSupportedFeatures` inside the
sync.Once and will assign a global featuresSupported variable with the
supported features after the call.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>