Add an lcow core dump cri-containerd testcase
This change adds a cri-containerd testcase that excercises the lcow core dump
functionality. It sets up two containers, one that uses an image that stackoverflows
shortly after starting. This should generate a core dump file in the sandbox mount location
in the test. Then the second container mounts the same sandbox mount and just verify that
the core dump file generated by the first is present.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>