Make ncproxy a urfave/cli app
This kind of started solely because the containerd version of the service flags
(because of urfave/cli) are double dashed, while flags from the flag package in
the stdlib are single, so writing generic code to launch both services just
by providing the name of the service was *mildly* annoying 😋.

After actually thinking about it though, there's probably a couple commands
down the road that might be helpful. One to start is to output a default config
for ncproxy for local testing or for actual production use which is in this
change. It will output a default config to stdout or to a file if the --file
flag is specified.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>