Add sleep before layer operation retries
This change adds a small sleep before a re-attempt on layer operation
failures. These failures should only happen on RS5 and the probable cause is because
of a different way in which container loopback vhds were mounted on this OS version.
A theory of why things might go awry on RS5 is due to some events from pnp getting reported
too late/early. If the prognosis is correct, a small sleep might help to try and get
things back into a "good" state before a reattempt.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>