Add GetCachedSupportedFeatures method to hcn package
To avoid a breaking change on GetSupportedFeatures introduce a new
GetCachedSupportedFeatures method. This method does the feature check
and version parsing once and then assigns a global with the information.
This can be used to optimize for situations where many uses of the
hcn.IsXSupported methods are going to be used (kube-proxy for example).

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit 112b5e7a30d0cac219085a26fc3a8188bc471db9)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>