Remove assigning the oci rtime working directory for job containers
All of our docs for the current way the filesystem layout works is that
relative paths will work, which *will* work if the working directory is set
to the scratch volume, however it generally won't be. All of the windows container
images have either C:\ or C:\windows\system32 set as the working directory, and
that's currently the only images usable.

We can look at documenting this for beta, but for now just use the behavior that
follows how the docs are worded.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>