Add os, os/exec and internal/syscall/execenv packages from stdlib
Due to some functionality in the stdlib not being supported and not wanting to
reimplement os/exec.Cmd and friends, we need to fork the stdlib. The functionality
being mentioned is related to not being able to do some setup with STARTUPINFOEXW
and exposing it in a go friendly way in the stdlib isn't really feasible as
there's currently no way to create a job object in the stdlib, and the syscall
package is locked down.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>