Get stdlib fork adjusted and working
This change edits a whole bunch of stuff in the stdlib fork to get things
actually building. This includes things such as removing the imports to
/internal packages in the stdlib (as we can't access them), changing around
the fields of some of the calls to point to our forked version instead of
the stdlib, and changing a whole slew of syscall definitions to x/sys/windows
equivalents. Also satisfying golangci-lint by ignoring some errors and removing
dead code that isn't used.

Additionally adds a small test to make sure it works.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>