Rework how working directory for job containers
Instead of taking the working directory as is, change to joining the working directory
requested with where the sandbox volume is located. It's expected that the default behavior
would be to treat all paths as relative to the volume as this would be equivalent to a
normal Windows Server Containers behavior.

For example:
A working directory of C:\ would become C:\C\12345678\
A working directory of C:\work\dir would become C:\C\12345678\work\dir

Signed-off-by: Daniel Canter <dcanter@microsoft.com>