Adjust execed processes OCI spec for LCOW
In most places in Containerd an execed process gets assigned the process spec of the
container it’s getting launched in (so it will inherit anything set on the container
spec), but due to the nature of LCOW there’s some OCI spec fields we set in the UVM
instead as they’re not able to be handled on the host (in a clean manner at least).
One of these is the user for the container. On a Linux host, Containerd will check if
the user exists in the filesystem for the container before setting the user on the spec.
On LCOW we just have vhd’s with the contents of the layers when making the spec which makes
this a bit infeasible, so we defer that work until we’re in the guest and then edit the spec
if the user exists. This has the outcome that the user is never set on the containers spec on
the host for LCOW, but we do have the final amended spec in the UVM with whatever user was
requested (or was set in the image via USER and so on).

This change adds a function to adjust the execed processes spec to contain the user from
the container, but also leaves it open to add future fields that we may have to amend in the guest.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>