Add security policy enforcement of environment variables
Supports two different matching schemes:

- string

This is a direct string match. All characters must be equal.

- re2

The rule is an re2 regular expression that will be matched against the environment variable.

Environment variables are in the form "KEY=VALUE" as a single string.

The securitypolicy tool has been updated to automatically include any environment variables defined
in the image spec for an image to the allowed environment variables in the generated policy.

Signed-off-by: Sean T. Allen <seanallen@microsoft.com>