test: Make main.go a test file
Previously main.go didn't have _test suffix, so it was not considered a
test file. Notably this meant that TestMain was never actually invoked
because it must be in a test file. It seems we were fortunate that there
was nothing in TestMain that wasn't done automatically by the Go test
infrastructure.

As the cri-containerd directory is all test code, it is probably safe
to rename all other files in the directory to be test code as well.
However, that is left for a future change.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>