test: Add containerd start/stop support
This change lets the cri-containerd tests start/stop containerd as
needed, rather than assuming it is always running. This is done through
the addition of startContainerd/stopContainerd functions which can be
called from tests. As all of the existing tests need containerd to be
running, this currently is not used in any tests. Future tests can take
advantage of this functionality.

Tests assume that containerd is running when they start, and should not
need to explicitly start containerd before calling into it. This means
that if a test stops containerd, it needs to ensure containerd is
started again. If containerd crashes during a test, then subsequent
tests will fail, but that's the same as the current behavior.

An unfortunate side effect of this change is that, due to a standing
issue with Go's service support and containerd, the service can
sometimes exit with ERROR_PROCESS_ABORTED when it is stopped. Combined
with the fact that recovery actions are used for containerd, this can
result in the service being restarted by the service control manager. To
work around this, we need to first disable recovery actions for the
service before running tests. This can be done with:

  sc failure containerd reset=0 actions= command=

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>