Enforce security policy at unmount
This is the first iteration of policy enforcement at unmount. There is an additional set
of functionality that will come as part of a larger change in the near future.

With this commit, we record that a device has been unmounted such that it isn't eligible to
be used in any overlay after unmounting.

In a future commit, I will be adding disallowing unmounting a device that is being used by a
running container.

Signed-off-by: Sean T. Allen <seanallen@microsoft.com>