Update naming in internal security policy tool
Maksim pointed out that when we added information beyond the image of an image
that the "image" entries in a TOML policy generation file weren't describing
images; the describe containers.

The addition of command line, environment variables, and what not to allow
is a description of a container that should be allowed to be created. The
only image specific bit is the name.

Signed-off-by: Sean T. Allen <seanallen@microsoft.com>