Change internal data structure in SecurityPolicyEnforcer
This commit changes the data structure that we use to track possible
GCS container ids for a given security policy container from an array
to a set.

Set is the correct data structure to represent our constraint of "id
should only appear once".

Additionally, it makes this tricky bit of code slightly easier to understand.

Signed-off-by: Sean T. Allen <seanallen@microsoft.com>