Add WCOW sandbox mount support
This change adds sandbox mount like support for WCOW. Sandboxmounts are our LCOW
solution similar to a k8s empty dir mount. We create a directory that will house
various other subdirectories that a user can specify by supplying a sandbox:// prefix
for the host path of a mount. In the usual case the hostpath of a mount is in the context of the
physical host (e.g. I want C:\path on my machine mapped to C:\containerpath in my container),
however for sandbox mounts the host path is treated as relative to this directory we have
made in the VM. The root directory for these sandbox mounts I defined as C:\SandboxMounts

Example:
"mounts": [
    {
    "container_path": "C:\\test",
    "host_path": "sandbox:///test"
    }
]

The above will make a directory at C:\SandboxMounts\test in the Utility VM that will be mapped to
C:\\test in the container. If another container in the same pod supplied the same mount then
you would end up "sharing" this directory with the other container, meaning you would
both see anything placed/modified in this directory.

The backing storage for these mounts will be the UVMs scratch disk, which currently is always 10GB
(8.5 actually usable) as that's whats hardcoded in HCS for the call we use that generates the vhd.
For some reason the expand vhd function from HCS doesn't function for the VM scratch disk which needs
some investigation :(

Signed-off-by: Daniel Canter <dcanter@microsoft.com>