Check in CI that the test/ module doesn't use internal APIs
An idea mentioned in-passing in #1073, this would ensure that tests can
be looked at as usage examples, which use of internal APIs prevents.

vendor/ and internal/ are excluded, the former because we only care
about direct dependencies, and the latter so that if tests are done that
depend on an unpublished API, it's possible _and_ explicit.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>