Add process dump cri-containerd testcases
This change adds cri-containerd testcases that excercises the process dump
functionality. It sets up two containers, using images that stackoverflow/
throw an exception shortly after starting. This should generate a dump
file in the sandbox mount location in the test. Then the second container
mounts the same sandbox mount and just verify that the core dump file generated
by the first is present.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>