Revert "Update windowsbuilds.go"
This reverts commit 7b6e3dac2b3b0c2e06ab5d0bf2a0665fb80ae511. The reason
for this is because the tag 21h2 actually refers to a couple different
Windows builds now unfortunately. It's the tag for the latest Windows 10
update, for Windows 11, and Windows Server 2022. We're looking into how
best to actually name these going forward if we're going to account for an
event like this again.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>