Add reconnect logic for stdio pipes
This change adds retry logic on the stdio relay if the server end of the named pipe
disconnects. This is a common case if containerd restarts for example.
The current approach is to make a io.Writer wrapper that handles the
reconnection logic on a write failure if it can be determined that the error
is from a disconnect.

This exposes a new shim config option to tailor the retry timeout.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>