Use cenkalti/backoff/v4 for io retry logic
This adds cenkalti/backoff/v4 as a dependency to be used for handling
exponential backoff logic for stdio connection retry attempts. Retrying
at a fixed interval is a bit naive as all of the shims would potentially
be trying to reconnect to 3 pipes continuously all in <timeout> bursts.
This allows us to space out the connections, set an upper limit on timeout
intervals and add an element of randomness to the retry attempts.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>