Fix LayerData not being usable for ComputeStorage package
Previously the LayerData structure in the computestorage package used
definitions from the hcs schema from /internal so it was not actually possible
to create a LayerData structure for an outside caller.

This just creates local type aliases for hcsschema.Version and hcsschema.Layer
so a client can create the structure now using computestorage.Version and
computestorage.Layer respectively.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>