Support booting isolated SNP from a GuestStateFile rather than separate kernel and userland (initrd.img)
Mainly this refactors the code that creates the hcs api json document into two paths. One is the previous
logic that will create a kernel command line and boot the kernel and userland from individual files using
the "LinuxKernelDirect" scheme. With isolation enabled this must be replaced with "GuestState/GuestStateFilePath"
etc to load from a vmgs file.

There are updates and addition files in the schema2 directory to support the newer hcs API so that the existing
way where an object that represents the hcs api json is built and then serialised to json can be used with
isolation_setting etc.

If a SecurityPolicy annotation is present it will boot the vmgs file unless
"io.microsoft.virtualmachine.lcow.no_security_hardware" is set to true. The various example pod.json files
will need to be updated for use with non SNP machines.

Signed-off-by: Ken Gordon <kegordo@microsoft.com>