Run Go 1.17 gofmt on repo
This change runs gofmt on Go 1.17 in /internal, /cmd, /hcn, and /test.

Go 1.17 added a new syntax for build tags that gofmt automatically applies.
The new format will just be ignored on older builds and 1.17 and up will
prefer the new syntax (which is much nicer if I may add), so this should
be harmless.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>