Add conpty (pseudo console) package
This change adds a conpty package that houses go friendly wrappers around
the Pseudo Console API in Windows. This will be used to support tty scenarios
for Host Process containers.

There's not many tests I can add here as you need to hook this up to a running
process, where that work is coming.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>