Set default time zone for WCOW UVM (#1192)
For the v2 hcs code paths it seems the only time a time zone is set is if
a new field on the guest connection settings is present (which we don't have)
while using the internal guest connection (shim -> hcs -> gcs). Otherwise
the guest is just left without a time zone set, so things like tzutil or
the get-timezone powershell cmdlet will return an invalid time zone set.
We swapped to always using the external guest connection we maintain in the
shim so we need to set a time zone explicitly.

This change issues a request to the gcs to set a timezone via the same method that
hcs uses internally. It sets the guests time zone to whatever is present on
the host which is the docker behavior, and then all containers in the vm
should inherit this. Additionally expose an option to override this behavior and
just set the time zone to UTC. If the container wants to change its time zone
to something else, it is free to on supported images.

See https://docs.microsoft.com/en-us/virtualization/windowscontainers/manage-containers/virtual-time-zone

Signed-off-by: Daniel Canter <dcanter@microsoft.com>