Add ws2022 image/build to cri-containerd tests (#1160)
* Add ws2022 image/build to cri-integration tests

This change adds a new case to the getWindowsServerCoreImage and
getWindowsNanoserverImage functions to return ws2022 on a ws2022 or higher
build. The higher case is because of some recent efforts to improve down-level
compatability for Windows container images. For reference, the ltsc2022 image
works on a win11 host without hypervisor isolation.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit f099e34878c260511ff37873b7231b5f2c769199)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>