Add new exec package for host process containers (#1233)
* Add new exec package for host process containers

This change adds a new exec package thats main goal is to run
external processes on Windows. Unfortunately due to a couple
things that can't be accomplished with the stdlib os/exec package,
this new package is meant to replace how processes for host process
containers are launched.

The main shortcomings are not being able to pass in a pseudo console to use
for tty scenarios, and not being able to start a process assigned to a job object
instead of doing the Create -> Assign dance. Both of these issue are centered
around not having access to the process thread attribute list that is setup inside
of syscall.StartProcess. This is needed to be able to properly setup both cases,
as it requires calling UpdateProcThreadAttribute and passing in what's necessary
for both scenarios.

This change ends up bumping x/sys/windows as well to grab some fixes for the attribute list functionality.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit 2314362e977aa03b3ed245a4beb12d00422af0e2)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>