Rename conpty.New to conpty.Create
This change renames conpty.New, the method used to create a new ConPTY
object, to Create instead. Mostly preference and to stay in line with what
we'd named the method for creating a job object. The windows API used to
create the pty is named 'CreatePseudoConsole' so to me it makes more sense.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>