Rework TestPseudoConsolePowershell (#1255)
Before this test used to write something to the pty and then see if we
could read the same thing shortly afterwards. Really all we wanna test here
is just that we can write things in general, so the same thing that
TestExecStdinPowershell is currently doing should suffice.

If the process exits then the 'exit' write went through so that should be
plenty for this test. This change additionally adds a timeout for waiting
for the process to exit for TestPseudoConsolePowershell as well as
TestExecStdinPowershell so we have some indicator for if the 'exit' write
didn't work.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>
(cherry picked from commit e093fbd404aeaa51f9728dab1125bd48b52f3d5a)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>