Rework process dump cri-containerd tests
This change aims to deflake the  Windows process dump test by continuously
checking for a .dmp instead of waiting some arbitrary amount of time and
then checking once. While we haven't seen any flakes with the Linux version
of this test for generating coredumps, the same logic is employed for the
LCOW test as well.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>