Add ErrInvalidHandle and fix list stats (#1276) (#1277)
When querying the stats of a container that is in the process of being
stopped, an ERROR_INVALID_HANDLE (0x6) may be returned. This change
ignores that error and returns an empty stats object.

This change also fixes the return values of Stats() when encountering
one of the expected errors. Returning nil stats when error is nil will
break caller assumptions of finding a valid value when error is nil.
Instead, an ErrNotFound is returned.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>
(cherry picked from commit 790bcae4c9e9734f4ca50d0e17b03ed749f2b4d9)
Signed-off-by: Daniel Canter <dcanter@microsoft.com>