Add local user account creation for host process containers
This change is experimental for now. This allows a user the ability to pass
a Windows group name as the username for the container. What happens in this
case is:

1. Client provides a Windows group name as the username for the container
2. We validate that it's a group, and if so then make a temporary local
account.
3. Add local account to the group passed in and run the container under the
account.
4. On container exit delete the account.

This allows a client to setup a group with whatever permissions/restrictions
needed on it and have any host process containers run as a user in the group.
The blocker for not just creating a user itself with the right permissions is
there's no cri field to pass a user password, and passwordless logon seems to be
blocked on Windows by default.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>