Expand env variables to job mount path
When installing tools into container it is sometimes desirable to have
those tools on the path for the current process.  Since the Windows
doesn't support variable expansion on the PATH variable and the
job mount path as something that isn't know at runtime we need to expand
it for the passer.  This could go away if the mount path is no longer
used in the future.

Signed-off-by: James Sturtevant <jsturtevant@gmail.com>