Cleanup 'getUserTokenInheritAnnotation' (#1294)
This function used to just return a bool of whether the annotation was "true"
or not which is a bit diffferent than what the name implied. Change to inheritUserTokenIsSet
to be more clear. It also used to check if the annotation was set internally which
shouldn't be needed as we only cared about whether it was "true" or not. If the
value isn't in the map we should just get back the default value of a string which
is fine for what it was being used for.

Additionally moves a comment that was mistakingly directly above the function body
to the top of the file.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>