Linux GCS: Log disk info on ENOSPC errors
We've had quite a few times where creating the upper or work directories in
the guest fails with ENOSPC but we don't have any view into what the mount
looks like at the time we get this. This change just catches any ENOSPC errors
when creating an overlayfs mount, calls statfs and logs the disk space and inode
info for the mount the failed directory is on. This should make investigating
these types of issues much easier.

This may be followed up with a change to delete the upper and work directories
for a container, as this becomes troublesome with the model for sharing a scratch volume.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>