Span sampling and export; logging context integration
Updated span export to:
* include span kind
* log if span attributes were dropped
* format time as string instead

Added `log.S()` to set the log entry stored in the context with
provided fields. `log.G()` now checks the context for a stored context.
Added `log.Copy()` to add log entry and trace span from source context
to destination, allowing for duplicating contexts but not cancellation.
Added `log.U()` to update the context an entry (in the context) points
to, allowing it to reference the latest span and other information.
Added `oc.StartSpan` to set the context log-entries reference to the
newly created valaue.

Added helper `log.Format*` functions to format Time and other structs
to JSON, but only if the logging level is high enough that the
information would be logged.

Set span kind to client/server, as required, for bridge RPC calls.

Updated internal/cmd to use spans within `.Start()` and `.Wait()`

Reduced the number of `INFO` level logs printed, downgrading to
`DEBUG`.

Moved traces and spans from inside `Once.Do()`, so that they are
always logged, even if the `Once` does not execute.

Signed-off-by: Hamza El-Saawy <hamzaelsaawy@microsoft.com>