Consolidate directory cleanup for overlayfs mounts
Previously after making the three directories necessary for the mount
(upper, work and the target) we would defer a function after each creation
to clean it up if we encountered an error anywhere after. This change just
moves all three of deferred os.RemoveAlls to one defer at the top of the function to
consolidate the logic.

os.RemoveAll if the path doesn't exist (e.g. the case where we only got around
to making one of the directories and something failed and we haven't made the
other two) just returns nil so this should be harmless.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>